Store and fetch scraping snapshots to Firestore
