+++ types.d.ts
@@ -1,6 +1,6 @@
 export interface UsedCar {
   image?: string;
-  link?: string;
+  link: string;
   make: string;
   model: string;
   modelExtra: string;
